+++ src/compiler/mod.rs
@@ -1,9 +1,8 @@
 mod ast_lexer;
-pub mod ast_parser;
-mod code_section_parser;
+pub(crate) mod ast_parser;
+pub(crate) mod code_section_parser;
 mod modifier_lexer;
 mod modifier_parser;
 pub mod spreadsheet;
 pub mod spreadsheet_cell;
-pub mod template;
-pub mod token_library;
+pub(crate) mod token_library;